+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 _No unreleased changes_
 
+## [1.0.0] - 2023-07-04
+### Added
+- Added UI
+- Added force for restore mode
+- Added checking for updates on launch
+
 ## [0.1.1] - 2023-07-03
 ### Fixes
 - Fixed typo in output message
@@ -14,6 +20,7 @@ _No unreleased changes_
 ## [0.1.0] - 2022-12-23
 ### Initial Release
 
+[1.0.0]: https://github.com/aosterwyk/elite-binds-backup-script/tree/v1.0.0
 [0.1.1]: https://github.com/aosterwyk/elite-binds-backup-script/tree/v0.1.1 
 [0.1.0]: https://github.com/aosterwyk/elite-binds-backup-script/tree/v0.1.0 
 [Unreleased]: https://github.com/aosterwyk/elite-binds-backup-script/compare/master...dev